Refactor AuthenticationTokenRepository to be easier to unit test.
