+++ app/src/main/java/com/davidread/habittracker/common/di/CommonModule.kt
@@ -1,8 +1,15 @@
 package com.davidread.habittracker.common.di
 
+import android.app.Application
 import com.davidread.habittracker.common.constant.BaseUrl
 import com.davidread.habittracker.common.network.AuthenticationInterceptor
 import com.davidread.habittracker.common.repository.AuthenticationTokenRepository
+import com.davidread.habittracker.common.repository.PREFS_NAME
+import com.google.crypto.tink.Aead
+import com.google.crypto.tink.KeysetHandle
+import com.google.crypto.tink.RegistryConfiguration
+import com.google.crypto.tink.aead.AeadKeyTemplates
+import com.google.crypto.tink.integration.android.AndroidKeysetManager
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -11,6 +18,9 @@ import okhttp3.OkHttpClient
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 import javax.inject.Singleton
+
+private const val KEYSET_NAME = "auth_keyset"
+private const val MASTER_KEY_URI = "android-keystore://auth_master_key"
 
 @Module
 @InstallIn(SingletonComponent::class)
@@ -38,5 +48,16 @@ class CommonModule {
     @Singleton
     fun providesAuthenticationInterceptor(authenticationTokenRepository: AuthenticationTokenRepository): AuthenticationInterceptor {
         return AuthenticationInterceptor(authenticationTokenRepository)
+    }
+
+    @Provides
+    fun providesAead(application: Application): Aead {
+        val keysetHandle: KeysetHandle = AndroidKeysetManager.Builder()
+            .withSharedPref(application, KEYSET_NAME, PREFS_NAME)
+            .withKeyTemplate(AeadKeyTemplates.AES256_GCM)
+            .withMasterKeyUri(MASTER_KEY_URI)
+            .build()
+            .keysetHandle
+        return keysetHandle.getPrimitive(RegistryConfiguration.get(), Aead::class.java)
     }
 }
