+++ app/src/main/java/com/davidread/habittracker/common/repository/AuthenticationTokenRepository.kt
@@ -2,37 +2,26 @@ package com.davidread.habittracker.common.repository
 
 import android.app.Application
 import android.content.Context
-import android.util.Base64
 import androidx.core.content.edit
 import com.davidread.habittracker.common.model.Result
+import com.davidread.habittracker.common.util.Base64
 import com.google.crypto.tink.Aead
-import com.google.crypto.tink.KeysetHandle
-import com.google.crypto.tink.RegistryConfiguration
-import com.google.crypto.tink.aead.AeadKeyTemplates
-import com.google.crypto.tink.integration.android.AndroidKeysetManager
 import javax.inject.Inject
 
-private const val PREFS_NAME = "secure_prefs"
-private const val KEYSET_NAME = "auth_keyset"
-private const val MASTER_KEY_URI = "android-keystore://auth_master_key"
-private const val TOKEN_KEY = "auth_token"
+internal const val PREFS_NAME = "secure_prefs"
+internal const val TOKEN_KEY = "auth_token"
 
-class AuthenticationTokenRepository @Inject constructor(private val application: Application) {
-
-    private val aead by lazy {
-        val keysetHandle: KeysetHandle = AndroidKeysetManager.Builder()
-            .withSharedPref(application, KEYSET_NAME, PREFS_NAME)
-            .withKeyTemplate(AeadKeyTemplates.AES256_GCM)
-            .withMasterKeyUri(MASTER_KEY_URI)
-            .build()
-            .keysetHandle
-        keysetHandle.getPrimitive(RegistryConfiguration.get(), Aead::class.java)
-    }
+class AuthenticationTokenRepository @Inject constructor(
+    private val aead: Aead,
+    private val base64: Base64,
+    private val application: Application
+) {
 
     fun saveAuthenticationToken(token: String): Result<Unit> {
         try {
             val encryptedToken = aead.encrypt(token.toByteArray(), null)
-            val encryptedTokenBase64 = Base64.encodeToString(encryptedToken, Base64.DEFAULT)
+            val encryptedTokenBase64 =
+                base64.encodeToString(encryptedToken, android.util.Base64.DEFAULT)
             application.getSharedPreferences(PREFS_NAME, Context.MODE_PRIVATE).edit {
                 putString(TOKEN_KEY, encryptedTokenBase64)
             }
@@ -48,7 +37,8 @@ class AuthenticationTokenRepository @Inject constructor(private val application:
                 application.getSharedPreferences(PREFS_NAME, Context.MODE_PRIVATE)
                     .getString(TOKEN_KEY, null)
                     ?: throw IllegalStateException("encryptedToken cannot be null")
-            val encryptedTokenBytes = Base64.decode(encryptedTokenBase64, Base64.DEFAULT)
+            val encryptedTokenBytes =
+                base64.decode(encryptedTokenBase64, android.util.Base64.DEFAULT)
             val decryptedToken = String(aead.decrypt(encryptedTokenBytes, null))
             return Result.Success(decryptedToken)
         } catch (e: Exception) {
@@ -56,6 +46,7 @@ class AuthenticationTokenRepository @Inject constructor(private val application:
         }
     }
 
+    // TODO: Remove this if it remains unused.
     fun clearAuthenticationToken() {
         application.getSharedPreferences(PREFS_NAME, Context.MODE_PRIVATE).edit {
             remove(TOKEN_KEY)
